Separated functions to prepare the TeX port for #23
